event , interested,profile,editProfile and API added

+++ lib/Features/Profille/Presentation/manager/cubits/deleteProfile_cubit/deleteProfile_cubit.dart
@@ -1,3 +1,4 @@
+import 'package:eventk/Core/dataBase/Cache/Cache_Helper.dart';
 import 'package:eventk/Core/errors/exceptions.dart';
 import 'package:eventk/Core/utils/delete_profile_service.dart';
 import 'package:eventk/Features/Profille/Presentation/manager/cubits/deleteProfile_cubit/deleteProfile_state.dart';
@@ -13,6 +14,10 @@ class DeleteprofileCubit extends Cubit<DeleteprofileStates> {
     emit(DeleteProfileLoadingState());
     try {
       final deleteProfileModel = await service.deleteProfile(oldPass);
+      if (deleteProfileModel.status != "Success") {
+        throw CustomExceptions(deleteProfileModel.message);
+      }
+      await CacheHelper().removeData(key: 'token');
       emit(DeleteProfileSuccessState(deleteProfileModel));
     } on CustomExceptions catch (e) {
       emit(DeleteProfileFailureState(errorMessage: e.message));
